Ignore the BaseView
