+++ app/resource/models/notes.py
@@ -86,6 +86,6 @@ class Note(Document):
 
     title = StringField()
     content = StringField(required=True)
-    meta = EmbeddedDocumentField(NoteMeta, default=NoteMeta)
+    # meta = EmbeddedDocumentField(NoteMeta, default=NoteMeta)
 
     user = ListField(ReferenceField(User, reverse_delete_rule=CASCADE))
